mobile/android/tests/browser/chrome/test_restricted_profiles.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1042715
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1042715</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript;version=1.7">

  const { classes: Cc, interfaces: Ci, utils: Cu } = Components;

  Cu.import("resource://gre/modules/ctypes.jsm");
  Cu.import("resource://gre/modules/JNI.jsm");
  Cu.import("resource://gre/modules/Task.jsm");

  function test_isUserRestricted() {
    // Make sure the parental controls service is available
    ok("@mozilla.org/parental-controls-service;1" in Cc);

    let pc = Cc["@mozilla.org/parental-controls-service;1"].createInstance(Ci.nsIParentalControlsService);

    // In an admin profile, like the tests: enabled = false
    // In a restricted profile: enabled = true
    ok(!pc.parentalControlsEnabled);
    ok(!pc.blockFileDownloadsEnabled);

    ok(pc.isAllowed(Ci.nsIParentalControlsService.DOWNLOAD));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.INSTALL_EXTENSION));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.INSTALL_APP));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.BROWSE));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.SHARE));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.BOOKMARK));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.INSTALL_EXTENSION));
    ok(pc.isAllowed(Ci.nsIParentalControlsService.MODIFY_ACCOUNTS));
  }

  test_isUserRestricted();

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1042715">Mozilla Bug 1042715</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testRestrictedProfiles</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
